chore(ui/dataLoaders): Move selection step components to dataloaders dir
